Add t-test stats
