+++ package.json
@@ -19,6 +19,7 @@
     "@typescript-eslint/parser": "2.12.0",
     "beautify-benchmark": "^0.2.4",
     "benchmark": "^2.1.4",
+    "cephes": "^1.2.0",
     "chance": "^1.1.3",
     "commonmark": "0.29.0",
     "commonmark-spec": "0.29.0",
@@ -42,6 +43,7 @@
     "simple-statistics": "7.0.7",
     "ts-jest": "24.2.0",
     "ts-loader": "6.2.0",
+    "ttest": "2.1.0",
     "typescript": "3.7.3",
     "uuid": "3.3.3"
   },
