+++ performance/src/lib/index.ts
@@ -1,1 +1,3 @@
 export * from "./profile";
+export * from "./t-test";
+export * from "./timing-stats";
